% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeEdges.R
\name{computeEdges}
\alias{computeEdges}
\title{Compute the layout of a Latent Class Tree}
\usage{
computeEdges(resTree)
}
\arguments{
\item{resTree}{A Latent Class Tree object}
}
\value{
Returns a matrix with in the first column the parent nodes (multiple entries for every split),
 the second column the child nodes and in the third column the level of the tree
}
\description{
Compute the layout of a Latent Class Tree
}
\details{
The \code{computeEdges} function constructs the layout of a LCT to be used by \code{igraph} and \code{qgraph}.
}

