% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeEdgesSplits.R
\name{computeEdgesSplits}
\alias{computeEdgesSplits}
\title{Compute the layout of a Latent Class Tree}
\usage{
computeEdgesSplits(resTree)
}
\arguments{
\item{resTree}{A Latent Class Tree object}
}
\value{
Returns a matrix with in the first column the parent nodes (multiple entries for every split),
 the second column the child nodes and in the third column the level of the tree.
}
\description{
Compute the layout of a Latent Class Tree
}
\details{
The \code{computeEdges} function constructs the layout of a LCT, but only the splits.
This can be used by \code{igraph} and \code{qgraph} to make a figure without redundant nodes.
}

