% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataCrayen.R
\docType{data}
\name{Crayen}
\alias{Crayen}
\title{Arabidopsis QTL data on gravitropism}
\format{An object of}
\source{
\href{http://qtlarchive.org/db/q?pg=projdetails&proj=moore_2013b}{QTL Archive}
}
\usage{
data(Crayen)
}
\description{
Data from a QTL experiment on gravitropism in
Arabidopsis, with data on 162 recombinant inbred lines (Ler x
Cvi). The outcome is the root tip angle (in degrees) at two-minute
increments over eight hours.
}
\references{
Moore et al. (2013) Genetics 195:1077-1086
(\href{http://www.ncbi.nlm.nih.gov/pubmed/23979570}{PubMed})
}
\keyword{datasets}

