% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExploreTree.R
\name{exploreTree}
\alias{exploreTree}
\title{Explore a Latent Class Tree model}
\usage{
exploreTree(res, syntaxThree = NULL, mLevelsCovariates = c(1, 2),
  method = "bch", dirTreeResults = getwd(), ResultsFolder = "exploreTree",
  Covariates = c("age", "sex"), mLevelsLGdependent = c("continuous",
  "ordinal"), weight = "weight")
}
\arguments{
\item{res}{Treeobject from}

\item{dirTreeResults}{Maximum size of the first split of the tree. Will be assessed with the criterion given in stopCriterium. Defaults to two.}

\item{ResultsFolder}{folder with results}

\item{LG}{Path to the Latent GOLD executable}

\item{levelsCovariates}{Vector with the levels of each of the covariates (1 = continuous, 1< = ordinal)}

\item{measurementLevels}{Whether the ordering of classes should be decreasing or not. Defaults to TRUE.}
}
\value{
None
}
\description{
Explore a Latent Class Tree model with Latent GOLD 5.1
}

