% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCGT.R
\name{LCGT}
\alias{LCGT}
\title{Illustration of crayon colors}
\usage{
LCGT(dataDir, LG, LGS = NULL, dependent = NULL, levelsDependent = 2,
  independent = NULL, caseid = NULL, weight = "weight",
  resultsName = "", maxClassSplit1 = 2, maxClassSplit2 = 2,
  decreasing = TRUE, stopCriterium = "BIC", minSampleSize = 5,
  nKeepVariables = 0, namesKeepVariables = NULL, sets = 16,
  iterations = 50)
}
\arguments{
\item{method2order}{method to order}

\item{cex}{character expansion for the text}

\item{mar}{margin paramaters; vector of len}
}
\value{
None
}
\description{
Illustration of crayon colors
}

