% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DichotomousTemp.R
\name{LCT}
\alias{LCT}
\title{Estimate a Latent Class Tree model}
\usage{
LCT(dataDir, LG, LGS, decreasing = TRUE, maxClassSplit1 = 2,
  maxClassSplit2 = 2, stopCriterium = "BIC", resultsName = "",
  minSampleSize = 5, itemNames = NULL, nKeepVariables = 0,
  weight = "weight")
}
\arguments{
\item{dataDir}{directory of the data or a dataframe with the data}

\item{LG}{directory of the Latent GOLD executable}

\item{LGS}{directory of Latent GOLD syntax for a model with 1- and 2-class splits}

\item{decreasing}{Whether the ordering of classes should be decreasing or not. Defaults to TRUE.}

\item{maxClassSplit1}{Maximum size of the first split of the tree. Will be assessed with the criterion given in stopCriterium. Defaults to two.}

\item{maxClassSplit2}{Maximum size of each split after the first split of the tree. Defaults to two.}

\item{stopCriterium}{Can be "LL", "AIC", or "BIC" (default)}

\item{resultsName}{name of a folder called to be made with all results by Latent GOLD}

\item{minSampleSize}{minimum sample size of a class. If this is below 1, a probability of the total sample size is used.}

\item{itemNames}{The names of the indicators. If this is not given, the rownames of the datafile will be used.}

\item{nKeepVariables}{Number of variables to be kept if one wants to explore the results with external variables.}

\item{weight}{Name of the variable with the weights. When all records are unique observations, this should be one for every observation.}
}
\value{
None
}
\description{
Estimate a Latent Class Tree model with Latent GOLD 5.1
}

