% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot1treeExpl.R
\name{plot1TreeExpl}
\alias{plot1TreeExpl}
\title{Plot the profile plots of a LCT in one image}
\usage{
plot1TreeExpl(resTree, explTree, rYlim = c(0, 1), mar = c(1, 1, 1, 1))
}
\arguments{
\item{resTree}{A Latent Class Tree object.}
}
\description{
1 pdf file with a growing tree.
}

