% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot1treeQ.R
\name{plot1treeQ}
\alias{plot1treeQ}
\title{Plot a growing tree}
\usage{
plot1treeQ(resTree, edg, dirLCTsplits, name = "Tree.pdf", vsize = 8,
  vsize2 = 0.5)
}
\arguments{
\item{resTree}{A Latent Class Tree object}

\item{edg}{Edges of the layout of the tree. Can be computed with \code{computeEdges}.}

\item{dirLCTsplits}{list of directories of an figure for every split. This is returned by \code{plotLCTsplits}.}

\item{name}{Name of the to be created pdf .}

\item{vsize}{heighth of the nodes}

\item{vsize2}{width of the nodes}
}
\description{
1 pdf file with a growing tree.
}

