% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotExplTreeCov.R
\name{plotExplTreeCov}
\alias{plotExplTreeCov}
\title{Plot figures with the mean or proportion of the covariates for every split.}
\usage{
plotExplTreeCov(explTree, ylab = "Probability of belonging to a class",
  xlab = "Value of the covariates", ylim = c(0, 1), xlim = NULL,
  lwd = 1, ...)
}
\arguments{
\item{explTree}{A Latent Class Tree to which the 3-step method is applied (with the function \code{exploreTree}).}
}
\description{
Several figures with the mean or proportion of the covariates for every split.
}

