% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLayoutLCT.R
\name{plotLayoutLCT}
\alias{plotLayoutLCT}
\title{Plot the layout of a Latent Class Tree}
\usage{
plotLayoutLCT(resTree, ...)
}
\arguments{
\item{resTree}{Results of a Latent Class Tree analysis in a LCT object}
}
\value{
The layout of a Latent Class Tree
}
\description{
Plot the layout of a Latent Class Tree
}

