% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotExplTree.R
\name{plotLCTexpl}
\alias{plotLCTexpl}
\title{Plot figures with the mean or proportion of the covariates for every split.}
\usage{
plotLCTexpl(explTree, ...)
}
\arguments{
\item{explTree}{A Latent Class Tree to which the 3-step method is applied (with the function \code{exploreTree}).}
}
\description{
Several figures with the mean or proportion of the covariates for every split.
}

