% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLCTlevels.R
\name{plotLCTLevels}
\alias{plotLCTLevels}
\title{Plot the class means or conditinal probabilities of every level of a LCT}
\usage{
plotLCTLevels(resTree, rYlim = c(0, 1), doPlot = FALSE, nameTree = "",
  dev = "png")
}
\arguments{
\item{resTree}{A Latent Class Tree object}

\item{rYlim}{the ylim of a plot of a continuous variable. By default for categorical variables c(0, 1) is used.}

\item{doPlot}{Logical, does the object need to be printed to a file as indicated by \code{dev}}

\item{nameTree}{if \code{doPlot} is TRUE (default), a folder called 'Levels'\code{nameTree} will be created.
In this folder of every level of the tree a plot is created, which is called \code{nameTree},
followed by the name of the split at hand.}

\item{dev}{device to create a plot, either "png", or "pdf"}
}
\description{
Produce the class means or conditinal probabilities of every split of a LCT
}

