% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printLCT.R
\name{print.LCT}
\alias{print.LCT}
\title{Print tree results to the terminal}
\usage{
\method{print}{LCT}(x, ...)
}
\arguments{
\item{dataDir}{directory of the data or a dataframe with the data}

\item{LG}{directory of the Latent GOLD executable}

\item{LGS}{directory of Latent GOLD syntax for a model with 1- and 2-class splits}

\item{decreasing}{Whether the ordering of classes should be decreasing or not. Defaults to TRUE.}

\item{maxClassSplit1}{Maximum size of the first split of the tree. Will be assessed with the criterion given in stopCriterium. Defaults to two.}

\item{maxClassSplit2}{Maximum size of each split after the first split of the tree. Defaults to two.}

\item{stopCriterium}{Criterium to decide on a split. Can be "LL" (logLikelihood), "AIC" or "BIC".}

\item{resultsName}{Name of a folder which will be created in the working directory and contains all results by Latent GOLD.}

\item{minSampleSize}{Minimum sample size of a class. If this is below 1, a probability of the total sample size is used.}

\item{itemNames}{The names of the indicators. If this is not given, the rownames of the datafile will be used.}

\item{nKeepVariables}{Number of variables to be kept if one wants to explore the results with external variables.}

\item{weight}{Name of the variable with the weights. When all records are unique observations, this should be one for every observation.}

\item{measurementLevels}{A character vector being either ordinal or continuous to indicate the measurement level of each variable. It is required when LGS is specified.}
}
\value{
None
}
\description{
These function display the results of a Latent Class Tree analysis
}

