% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relImpr.R
\name{relImpr}
\alias{relImpr}
\title{Compute relative improvement at the root of the tree}
\usage{
relImpr(x, criterion = "LL")
}
\arguments{
\item{x}{treeobject}

\item{Criterion}{relative improvement for which criterion, default is the log likelihood, other options are AIC and BIC}
}
\value{
None
}
\description{
Compute relative improvement by adding classes at the root of the tree
}

