% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSocialCapital.R
\docType{data}
\name{SocialCapital}
\alias{SocialCapital}
\title{Dataset on Social Capital}
\format{Every row of the dataset represents an unique respondent, while column represent a variable of the General Social Survey.}
\usage{
data(SocialCapital)
}
\description{
This is a data set on social capital from the General Social Survey. It has been originally used by Owen and Videras (2009).
}
\references{
Ann L. Owen and Julio Videras (2009) Reconsidering social capital: a latent class approach.
Empirical Economics, 37(3), 555-582.
(\href{https://link.springer.com/article/10.1007/s00181-008-0246-6}{Springer})
}
\keyword{datasets}

