% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExploreTree.R
\name{exploreTree}
\alias{exploreTree}
\title{3-step method Latent Class Trees}
\usage{
exploreTree(resTree, LG = LG, syntaxExpl = NULL, sizeMlevels = c(2, 1),
  method = "bch", dirTreeResults = getwd(), ResultsFolder = "exploreTree",
  Covariates = c("age", "sex"), mLevels = c("ordinal", "continuous"),
  weight = "weight")
}
\arguments{
\item{resTree}{A Latent Class Tree object}

\item{LG}{Path to the Latent GOLD executable}

\item{syntaxExpl}{Directory of Latent GOLD syntax for a model for the 3-step analysis. This can be used for more complex models, but}

\item{dirTreeResults}{Maximum size of the first split of the tree. Will be assessed with the criterion given in stopCriterium. Defaults to two.}

\item{ResultsFolder}{folder with results}

\item{measurementLevels}{Whether the ordering of classes should be decreasing or not. Defaults to TRUE.}
}
\value{
None
}
\description{
Compute the relation of some covariates with each class by use of the 3-step method.
}

