% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCT.R
\name{LCT}
\alias{LCT}
\title{Estimate a Latent Class Tree model}
\usage{
LCT(dataDir, LG, LGS = NULL, itemNames = NULL, mLevels = NULL,
  weight = "weight", resultsName = "", maxClassSplit1 = 2,
  maxClassSplit2 = 2, decreasing = TRUE, stopCriterium = "BIC",
  minSampleSize = 5, nKeepVariables = 0, namesKeepVariables = NULL,
  sets = 16, iterations = 50)
}
\arguments{
\item{dataDir}{directory of the data or a dataframe with the data}

\item{LG}{directory of the Latent GOLD executable}

\item{LGS}{directory of Latent GOLD syntax for a model with 1- and 2-class splits}

\item{itemNames}{The names of the indicators. If this is not given, the rownames of the datafile will be used.}

\item{weight}{Name of the variable with the weights. When all records are unique observations, this should be one for every observation.}

\item{resultsName}{Name of a folder which will be created in the working directory and contains all results by Latent GOLD.}

\item{maxClassSplit1}{Maximum size of the first split of the tree. Will be assessed with the criterion given in stopCriterium. Defaults to two.}

\item{maxClassSplit2}{Maximum size of each split after the first split of the tree. Defaults to two.}

\item{decreasing}{Whether the ordering of classes should be decreasing or not. Defaults to TRUE.}

\item{stopCriterium}{Criterium to decide on a split. Can be "LL" (logLikelihood), "AIC" or "BIC".}

\item{minSampleSize}{Minimum sample size of a class. If this is below 1, a probability of the total sample size is used.}

\item{nKeepVariables}{Number of variables to be kept if one wants to explore the results with external variables.}

\item{measurementLevels}{A character vector being either ordinal or continuous to indicate the measurement level of each variable. It is required when LGS is specified.}
}
\value{
None
}
\description{
Estimate a Latent Class Tree model with Latent GOLD 5.1
}

